[fix] 방문 기록 저장 방식을 DB에서 로그로 변경
