+++ src/main/java/earlybird/earlybird/log/visit/service/VisitEventLogServiceV2.java
@@ -1,0 +1,22 @@
+package earlybird.earlybird.log.visit.service;
+
+import earlybird.earlybird.log.util.LogUtil;
+import earlybird.earlybird.log.visit.service.request.VisitEventLoggingServiceRequest;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Primary;
+import org.springframework.stereotype.Service;
+
+import java.util.Map;
+
+@Slf4j
+@Primary
+@Service
+public class VisitEventLogServiceV2 implements VisitEventLogService {
+    @Override
+    public void create(VisitEventLoggingServiceRequest request) {
+        LogUtil.infoLog(
+                Map.of("client-id", request.getClientId(), "event-type", "client-visit"),
+                "visit log: client-id={}", request.getClientId());
+
+    }
+}
